add setup-manager to setup the api
