+++ api/docker-compose.yml
@@ -24,7 +24,7 @@ services:
       DATABASE_URL: ${DATABASE_URL}
       SERVER_SECRET: ${SERVER_SECRET}
     restart: on-failure
-    command: bash -c "node dist/index.js seed && node dist/index.js keygen"
+    command: node dist/index.js setup-api
     depends_on: 
       - mongo
