+++ api/package.json
@@ -11,11 +11,9 @@
     "test": "jest --detectOpenHandles",
     "test-watch": "npm run test -- --watchAll",
     "start-node": "node dist/index.js server",
-    "start-keygen": "node dist/index.js keygen",
-    "start-seed": "node dist/index.js seed",
+    "start-setup-api": "node dist/index.js setup-api",
     "start": "npm-run-all build start-node",
-    "keygen": "npm-run-all build start-keygen",
-    "seed": "npm-run-all build start-seed",
+    "setup-api": "npm-run-all build start-setup-api",
     "build-watch": "tsc --watch ",
     "serve-nodemon": "nodemon ./dist/index server",
     "serve": "run-p build-watch serve-nodemon",
